feat: add current_version.py
