+++ scripts/current_version.py
@@ -1,0 +1,9 @@
+from lib.version import read_current_version
+
+
+def main():
+    print(read_current_version())
+
+
+if __name__ == "__main__":
+    main()
